requirements.txt
jinja2
weasyprint
pyyaml
lxml
factur-xreportlab
